+++ chatlearn/schedule/model_manager.py
@@ -26,7 +26,7 @@ from chatlearn.data.storage import Storage
 from chatlearn.launcher import dlc_utils
 from chatlearn import FSDPModule
 from chatlearn.models.torch_module import TorchModule
-from chatlearn.models.vllm_module_v2 import VLLMModuleV2
+from chatlearn.models.vllm_module import VLLMModule
 from chatlearn.runtime.decorator import decorate_class_func
 from chatlearn.runtime.decorator import timeit, preprocess_compute, monitor_error
 from chatlearn.runtime.dist_actor import DistActor, DistTorchActor, DistVLLMActor, DistModel
@@ -270,7 +270,7 @@ class ModelManager:
         model.finalize()
 
         def actor_type():
-            if isinstance(model, VLLMModuleV2):
+            if isinstance(model, VLLMModule):
                 return DistVLLMActor
             if isinstance(model, TorchModule):
                 return DistTorchActor
@@ -400,8 +400,8 @@ class ModelManager:
             group = i // self.resouce_manager.gpu_per_node
             for replica in replicas:
                 num_gpus = 1.0 / len(replicas)
-                # if isinstance(replica.model, VLLMModuleV2) and replica.vllm_engine is None:
-                if isinstance(replica.model, VLLMModuleV2) and replica.should_create_engine_actor():
+                # if isinstance(replica.model, VLLMModule) and replica.vllm_engine is None:
+                if isinstance(replica.model, VLLMModule) and replica.should_create_engine_actor():
                     num_gpus = num_gpus / 2
                     replica.create_engine_actor(num_gpus, placement_group, group)
                     # we do not want to add engine actor to all_actors
